// src/app/models/cadastro-api.ts
export interface CadastroApi {
  id?: number;
  nomePessoal: string;
  cpf: string;
  telefone: string;
  email: string;
  nomeAcademia: string;
  cnpj: string;
  endereco: string;
  bairro: string;
  uf: string;
  complemento: string;
  cidade: string;
  plano: string;
  login: string;
  senha: string;
}